Shift pipeline specific functions into static methods & specify caching/verbosity via class attributes
